Seperated commands into their own files. Updated references.
